label.float()

+++ mmdet/models/losses/balanced_bce_loss.py
@@ -31,7 +31,7 @@ def balanced_bce_with_logits_loss(pred, label,
                                   reduction='mean'):
     assert reduction in ['mean', 'sum'], "'none' mode not supported in this loss."
     flattened_pred = pred.flatten()
-    flattened_label = label.flatten()
+    flattened_label = label.flatten().float()
     pos_inds = flattened_label.nonzero().reshape(-1).cpu()
     neg_inds = (flattened_label == 0).nonzero().reshape(-1).cpu()
     rand_neg_num = max(
